+++ feature-ui-exam-result/build.gradle.kts
@@ -27,6 +27,7 @@ dependencies {
         projects.utilKotlin,
         projects.utilCompose,
         projects.sharedUiCompose,
+        projects.utilExceptionHandling,
         libs.orbit.viewmodel,
         libs.orbit.compose,
         libs.ktx.lifecycle.runtime,
