+++ feature-ui-exam-result/src/main/java/team/duckie/app/android/feature/ui/exam/result/ExamResultActivity.kt
@@ -8,19 +8,26 @@
 package team.duckie.app.android.feature.ui.exam.result
 
 import android.os.Bundle
+import androidx.activity.compose.BackHandler
 import androidx.activity.compose.setContent
 import androidx.activity.viewModels
 import dagger.hilt.android.AndroidEntryPoint
+import org.orbitmvi.orbit.viewmodel.observe
 import team.duckie.app.android.feature.ui.exam.result.screen.ExamResultScreen
+import team.duckie.app.android.feature.ui.exam.result.viewmodel.ExamResultSideEffect
 import team.duckie.app.android.feature.ui.exam.result.viewmodel.ExamResultViewModel
 import team.duckie.app.android.navigator.feature.home.HomeNavigator
+import team.duckie.app.android.util.compose.ToastWrapper
+import team.duckie.app.android.util.exception.handling.reporter.reportToCrashlyticsIfNeeded
+import team.duckie.app.android.util.exception.handling.reporter.reportToToast
 import team.duckie.app.android.util.ui.BaseActivity
+import team.duckie.app.android.util.ui.finishWithAnimation
 import team.duckie.quackquack.ui.theme.QuackTheme
 import javax.inject.Inject
 
 @AndroidEntryPoint
 class ExamResultActivity : BaseActivity() {
-    val viewModel: ExamResultViewModel by viewModels()
+    private val viewModel: ExamResultViewModel by viewModels()
 
     @Inject
     lateinit var homeNavigator: HomeNavigator
@@ -28,12 +35,37 @@ class ExamResultActivity : BaseActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContent {
+            BackHandler {
+                finishWithAnimation()
+            }
+
             QuackTheme {
-                ExamResultScreen(
-                    navigateToHome = {
-                        homeNavigator.navigateFrom(this)
-                    },
-                )
+                ExamResultScreen()
+            }
+
+            viewModel.observe(
+                lifecycleOwner = this,
+                sideEffect = ::handleSideEffect,
+            )
+        }
+    }
+
+    private fun handleSideEffect(sideEffect: ExamResultSideEffect) {
+        when (sideEffect) {
+            is ExamResultSideEffect.ReportError -> {
+                sideEffect.exception.run {
+                    printStackTrace()
+                    reportToToast()
+                    reportToCrashlyticsIfNeeded()
+                }
+            }
+
+            ExamResultSideEffect.FinishExamResult -> {
+                finishWithAnimation()
+            }
+
+            is ExamResultSideEffect.SendToast -> {
+                ToastWrapper(this).invoke(message = sideEffect.message)
             }
         }
     }
