+++ feature-ui-exam-result/src/main/res/values/strings.xml
@@ -8,4 +8,5 @@
 <resources>
     <string name="solve_retry">다시 풀기</string>
     <string name="exit_exam">시험 끝내기</string>
+    <string name="feature_prepare">준비 중인 기능입니다.</string>
 </resources>
